--- packages/graphql-proxy/example/resolvers.js
@@ -1,1 +1,0 @@
-{"Query":{},"Mutation":{}}
